Rework the ignorefrom/ignoreto usage.
Previously, tracks outside the active range would be completely omitted. Now,
tracks prior to it will still be built (which takes no appreciable time if they
already existed), but tracks afterwards will not; the significance of the
ignorefrom point is that that's where they begin to be added to tracklist.

Also fixes, fwiw, an off-by-one error in ignorefrom. Since that value is an
estimate anyway, it's not an issue :)